forward port signedness fixes from RELENG_6 fix compile error for case where MSI_SUPPORTED not defined
Approved by: re (blanket)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@171868 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/dev/cxgb/cxgb_offload.c
@@ -596,7 +596,7 @@ cxgb_remove_tid(struct toedev *tdev, void *ctx, unsigned int tid)
 
 	BUG_ON(tid >= t->ntids);
 	if (tdev->type == T3A)
-		atomic_cmpset_ptr((long *)&t->tid_tab[tid].ctx, (long)NULL, (long)ctx);
+		atomic_cmpset_ptr((uintptr_t *)&t->tid_tab[tid].ctx, (long)NULL, (long)ctx);
 	else {
 		struct mbuf *m;
